edited pool

+++ src/main/scala/objektwerks/Entity.scala
@@ -30,7 +30,7 @@ final case class Account(id: Long :| Greater[0],
 final case class Pool(id: Long :| Greater[0],
                       accountId: Long :| Greater[0],
                       name: String :| MinLength[3], 
-                      built: Long :| Greater[0],
+                      built: Int :| Greater[0],
                       volume: Int :| GreaterEqual[100],
                       unit: UnitOfMeasure) extends Entity
